# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: FUKAMU Notes HTTP contract
  version: 0.1.0-go-migration
  description: >-
    External wire contract captured from main at
    f423da9932163980485ecc5bc2055b7c8c3b3d8b. Route availability describes
    the current production entry point; x-handler-contract records an
    implemented but currently disconnected handler and does not publish it.
servers:
  - url: /
paths:
  /api/launch-status:
    get:
      operationId: getLaunchStatus
      x-fukamu-state: connected
      responses:
        '200':
          description: Current fail-closed launch admission decision.
          headers:
            Cache-Control:
              { schema: { type: string, enum: [private, no-store] } }
            X-Content-Type-Options:
              { schema: { type: string, enum: [nosniff] } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LaunchStatus' }
        '503': { $ref: '#/components/responses/LaunchGateUnavailable' }
  /api/sync:
    post:
      operationId: legacySync
      x-fukamu-state: connected-in-legacy-test
      x-go-migration-state: connected-local-private-runtime
      description: >-
        The Go migration path additionally requires a verified configured
        legacy owner and exact same-origin request before reading the body.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SyncRequest' }
      responses:
        '200':
          description: Full legacy synchronization result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SyncResponse' }
        '400': { $ref: '#/components/responses/InvalidSyncRequest' }
        '403': { $ref: '#/components/responses/LaunchForbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413': { $ref: '#/components/responses/SyncPayloadTooLarge' }
        '500': { $ref: '#/components/responses/SyncFailed' }
        '503': { $ref: '#/components/responses/LaunchGateUnavailable' }
  /api/v2/sync:
    post:
      operationId: syncV2Route
      x-fukamu-state: disconnected
      x-handler-contract: sync-v2
      x-go-handler: backend/internal/httpapi/sync_v2.go
      description: >-
        The current route remains closed (404 in legacy-test, otherwise 503).
        The 200/400/401/402/403/409/413 Go handler contract is tested
        separately with PostgreSQL, quota, and encrypted-object adapters and
        is deliberately absent from the public HandlerOptions composition.
      responses:
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /api/billing/checkout:
    get:
      operationId: billingCheckoutOfferRoute
      x-fukamu-state: connected-local-fixture-production-closed
      x-handler-contract: billing-checkout-offer
      x-go-handler: backend/internal/httpapi/legal.go
      description: >-
        The guarded local-fixture profile serves the typed no-charge offer.
        The production/default route remains closed because it supplies no LegalRuntime.
      responses:
        '200':
          description: Current local fixture offer and canonical hash.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContractOfferResponse' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
    post:
      operationId: billingCheckoutRoute
      x-fukamu-state: connected-local-fixture-production-closed
      x-handler-contract: billing-checkout-submit
      x-go-handler: backend/internal/httpapi/legal.go
      description: >-
        The local fixture records immutable evidence and returns URL-free
        confirmation without charging or provider I/O. Production remains closed.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContractCheckoutRequest' }
      responses:
        '200':
          description: Provider redirect or URL-free local confirmation.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContractCheckoutResponse' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/RequestTooLarge' }
        '422': { $ref: '#/components/responses/InvalidRequest' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /api/billing/cancel:
    post:
      operationId: billingCancellationRoute
      x-fukamu-state: connected-local-fixture-production-closed
      x-handler-contract: billing-cancellation-period-end
      x-go-handler: backend/internal/httpapi/billing_cancellation.go
      description: >-
        The guarded local fixture returns a deterministic no-write period-end
        schedule. Production remains closed and Stripe is not constructed.
        Account deletion continues to use its distinct immediate-only port.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BillingCancellationRequest' }
      responses:
        '200':
          description: Scheduled or already-cancelled confirmation.
          content:
            application/json:
              schema:
                { $ref: '#/components/schemas/BillingCancellationResponse' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/RequestTooLarge' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /api/account/deletion:
    post:
      operationId: accountDeletionStartRoute
      x-fukamu-state: disconnected
      x-handler-contract: account-deletion-start
      x-go-handler: backend/internal/httpapi/account_deletion.go
      description: >-
        The production route remains closed. The authenticated Go contract
        stores the operation and capability before any deletion effect runs.
      responses:
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /api/account/deletion/status:
    post:
      operationId: accountDeletionResumeRoute
      x-fukamu-state: disconnected
      x-handler-contract: account-deletion-resume-one-step
      x-go-handler: backend/internal/httpapi/account_deletion.go
      description: >-
        The production route remains closed. The same-origin Go continuation
        contract consumes a capability sequence before at most one effect and
        does not require the session that the first effect revokes.
      responses:
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /api/account/privacy-requests:
    post:
      operationId: privacyRequestSubmitRoute
      x-fukamu-state: disconnected
      x-handler-contract: privacy-request-submit
      x-go-handler: backend/internal/httpapi/privacy_request.go
      description: >-
        The production route remains closed. The authenticated Go contract
        handler is tested separately and requires an explicit provider-neutral
        privacy runtime composition.
      responses:
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /api/account/privacy-requests/status:
    post:
      operationId: privacyRequestStatusRoute
      x-fukamu-state: disconnected
      x-handler-contract: privacy-request-status
      x-go-handler: backend/internal/httpapi/privacy_request.go
      description: >-
        The production route remains closed. The Go handler derives the owner
        from the secure session and never accepts Account or Vault identifiers.
      responses:
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /api/account/terms-consent:
    get:
      operationId: termsConsentStatusRoute
      x-fukamu-state: connected-local-fixture-production-closed
      x-handler-contract: terms-consent-status
      x-go-handler: backend/internal/httpapi/legal.go
      description: >-
        The guarded local fixture returns owner-scoped PostgreSQL consent
        status. Production/default remains closed without a LegalRuntime.
      responses:
        '200':
          description: Current owner-scoped terms consent status.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TermsConsentResponse' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
    post:
      operationId: termsConsentSubmitRoute
      x-fukamu-state: connected-local-fixture-production-closed
      x-handler-contract: terms-consent-submit
      x-go-handler: backend/internal/httpapi/legal.go
      description: >-
        The local fixture strictly decodes a bounded command after session and
        same-origin checks, then records/replays immutable evidence. Production is closed.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TermsConsentRequest' }
      responses:
        '200':
          description: Recorded or replayed owner-scoped consent.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TermsConsentResponse' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/RequestTooLarge' }
        '422': { $ref: '#/components/responses/InvalidRequest' }
        '503': { $ref: '#/components/responses/Unavailable' }
components:
  responses:
    LaunchGateUnavailable:
      description: Launch admission could not be established.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: launch-gate-unavailable }
    LaunchForbidden:
      description: The authenticated subject is not admitted.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Closed or unavailable in the current service mode.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: not-found }
    Unavailable:
      description: Closed or unavailable in the current service mode.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: unavailable }
    AuthenticationRequired:
      description: An authenticated local fixture session is required.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: authentication-required }
    Forbidden:
      description: Owner or same-origin verification failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: forbidden }
    InvalidRequest:
      description: Strict request decoding or an application precondition failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: invalid-request }
    Conflict:
      description: Stale disclosure, identifier conflict, or unavailable cancellation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RequestTooLarge:
      description: More than 2,048 request bytes.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: request-too-large }
    InvalidSyncRequest:
      description: Legacy request JSON or invariants are invalid.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 同期データが正しくありません。 }
    SyncPayloadTooLarge:
      description: More than 4,000,000 request bytes.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 同期データが正しくありません。 }
    SyncFailed:
      description: Legacy storage or synchronization failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: { type: string }
    LaunchStatus:
      type: object
      additionalProperties: false
      required: [publicAccessEnabled, userAllowed, canAccess, authenticated]
      properties:
        publicAccessEnabled: { type: boolean }
        userAllowed: { type: boolean }
        canAccess: { type: boolean }
        authenticated: { type: boolean }
    UuidV7:
      type: string
      minLength: 36
      maxLength: 36
      pattern: '^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-7[0-9A-Fa-f]{3}-[89ABab][0-9A-Fa-f]{3}-[0-9A-Fa-f]{12}$'
...
    ContractOfferResponse:
      type: object
      additionalProperties: false
      required: [offer, offerHash]
      properties:
        offer: { $ref: '#/components/schemas/ContractOffer' }
        offerHash: { $ref: '#/components/schemas/Sha256Digest' }
    ContractCheckoutRequest:
      type: object
      additionalProperties: false
      required: [submissionId, presentedOfferHash, consent]
      properties:
        submissionId: { $ref: '#/components/schemas/UuidV7' }
        presentedOfferHash: { $ref: '#/components/schemas/Sha256Digest' }
        consent: { $ref: '#/components/schemas/AffirmedConsent' }
    ContractCheckoutResponse:
      oneOf:
        - type: object
          additionalProperties: false
          required:
            [
              kind,
              evidenceOutcome,
              evidenceId,
              offerHash,
              offerVersion,
              checkoutUrl,
            ]
          properties:
            kind: { type: string, enum: [redirect] }
            evidenceOutcome: { type: string, enum: [recorded, replayed] }
            evidenceId: { $ref: '#/components/schemas/UuidV7' }
            offerHash: { $ref: '#/components/schemas/Sha256Digest' }
            offerVersion:
              { type: string, pattern: '^legal-commerce-v1:\d{4}-\d{2}-\d{2}$' }
            checkoutUrl: { type: string, format: uri, maxLength: 2048 }
        - type: object
          additionalProperties: false
          required: [kind, evidenceOutcome, evidenceId, offerHash, offerVersion]
          properties:
            kind: { type: string, enum: [local-confirmed] }
            evidenceOutcome: { type: string, enum: [recorded, replayed] }
            evidenceId: { $ref: '#/components/schemas/UuidV7' }
            offerHash: { $ref: '#/components/schemas/Sha256Digest' }
            offerVersion:
              { type: string, pattern: '^legal-commerce-v1:\d{4}-\d{2}-\d{2}$' }
    BillingCancellationRequest:
      type: object
      additionalProperties: false
      required: [idempotencyKey]
      properties:
        idempotencyKey:
          type: string
          minLength: 1
          maxLength: 255
          pattern: '^[A-Za-z0-9][A-Za-z0-9:_-]{0,254}$'
    BillingCancellationResponse:
      oneOf:
        - type: object
          additionalProperties: false
          required: [status, outcome, confirmedAt, accessEndsAt]
          properties:
            status: { type: string, enum: [cancellation-scheduled] }
            outcome: { type: string, enum: [scheduled] }
            confirmedAt: { $ref: '#/components/schemas/SafeInteger' }
            accessEndsAt: { $ref: '#/components/schemas/SafeInteger' }
        - type: object
          additionalProperties: false
          required: [status, outcome, confirmedAt, accessEndsAt]
          properties:
            status: { type: string, enum: [cancelled] }
            outcome: { type: string, enum: [already-cancelled] }
            confirmedAt: { $ref: '#/components/schemas/SafeInteger' }
            accessEndsAt: { $ref: '#/components/schemas/SafeInteger' }
...
    TermsConsentResponse:
      type: object
      additionalProperties: false
      required: [outcome, status]
      properties:
        outcome: { type: string, enum: [status, recorded, replayed] }
        status: { $ref: '#/components/schemas/TermsConsentStatus' }
    TermsConsentRequest:
      type: object
      additionalProperties: false
      required:
        [submissionId, presentedTermsVersion, presentedTermsHash, consent]
      properties:
        submissionId: { $ref: '#/components/schemas/UuidV7' }
        presentedTermsVersion:
          { type: string, pattern: '^terms-v1:\d{4}-\d{2}-\d{2}$' }
        presentedTermsHash: { $ref: '#/components/schemas/Sha256Digest' }
        consent: { $ref: '#/components/schemas/AffirmedConsent' }
...
    Mutation:
      type: object
      additionalProperties: false
      required:
        - mutationId
        - cardId
        - baseServerRevision
        - title
        - body
        - createdAt
        - updatedAt
        - kind
        - conflictIds
      properties:
        mutationId: { $ref: '#/components/schemas/UuidV7' }
        cardId: { $ref: '#/components/schemas/UuidV7' }
        baseServerRevision:
          nullable: true
          allOf: [{ $ref: '#/components/schemas/SafeInteger' }]
        title: { type: string, maxLength: 10000 }
        body:
          type: array
          maxItems: 10000
          items: { $ref: '#/components/schemas/BodySegment' }
        createdAt: { $ref: '#/components/schemas/SafeInteger' }
        updatedAt: { $ref: '#/components/schemas/SafeInteger' }
        kind: { type: string, enum: [upsert, resolve] }
        conflictIds:
          type: array
          maxItems: 500
          uniqueItems: true
          items: { $ref: '#/components/schemas/UuidV7' }
    SyncRequest:
      type: object
      additionalProperties: false
      required: [deviceId, mutations]
      properties:
        deviceId: { $ref: '#/components/schemas/UuidV7' }
        mutations:
          type: array
          maxItems: 500
          items: { $ref: '#/components/schemas/Mutation' }
    ServerCard:
      type: object
      additionalProperties: false
      required:
        [id, officialDisplayId, title, body, createdAt, updatedAt, revision]
      properties:
        id: { $ref: '#/components/schemas/UuidV7' }
        officialDisplayId:
          { type: integer, minimum: 1, maximum: 9007199254740991 }
        title: { type: string, maxLength: 10000 }
        body:
          type: array
          maxItems: 10000
          items: { $ref: '#/components/schemas/BodySegment' }
        createdAt: { $ref: '#/components/schemas/SafeInteger' }
        updatedAt: { $ref: '#/components/schemas/SafeInteger' }
        revision: { type: integer, minimum: 1, maximum: 9007199254740991 }
    Conflict:
      type: object
      additionalProperties: false
      required:
        [
          id,
          cardId,
          serverRevision,
          localTitle,
          localBody,
          serverTitle,
          serverBody,
          createdAt,
        ]
      properties:
        id: { $ref: '#/components/schemas/UuidV7' }
        cardId: { $ref: '#/components/schemas/UuidV7' }
        serverRevision: { type: integer, minimum: 1, maximum: 9007199254740991 }
        localTitle: { type: string, maxLength: 10000 }
        localBody:
          type: array
          items: { $ref: '#/components/schemas/BodySegment' }
        serverTitle: { type: string, maxLength: 10000 }
        serverBody:
          type: array
          items: { $ref: '#/components/schemas/BodySegment' }
        createdAt: { $ref: '#/components/schemas/SafeInteger' }
    SyncResponse:
      type: object
      additionalProperties: false
      required: [cards, conflicts, acknowledgedMutationIds]
      properties:
        cards:
          type: array
          maxItems: 100000
          items: { $ref: '#/components/schemas/ServerCard' }
        conflicts:
          type: array
          maxItems: 100000
          items: { $ref: '#/components/schemas/Conflict' }
        acknowledgedMutationIds:
          type: array
          maxItems: 500
          items: { $ref: '#/components/schemas/UuidV7' }